Move SierraGenerators into the "generators" namespace we use elsewhere
