+++ pipeline/transformer/transformer_sierra/src/test/scala/uk/ac/wellcome/platform/transformer/sierra/services/SierraTransformerWorkerTest.scala
@@ -10,13 +10,13 @@ import uk.ac.wellcome.models.work.internal.{IdentifierType, Work, WorkState}
 import uk.ac.wellcome.pipeline_storage.{PipelineStorageStream, Retriever}
 import uk.ac.wellcome.sierra_adapter.model.{
   SierraBibNumber,
-  SierraGenerators,
   SierraTransformable
 }
 import uk.ac.wellcome.sierra_adapter.model.Implicits._
 import uk.ac.wellcome.storage.generators.S3ObjectLocationGenerators
 import uk.ac.wellcome.storage.s3.S3ObjectLocation
 import uk.ac.wellcome.storage.store.memory.MemoryTypedStore
+import weco.catalogue.sierra_adapter.generators.SierraGenerators
 import weco.catalogue.source_model.SierraSourcePayload
 import weco.catalogue.transformer.{
   TransformerWorker,
