+++ sierra_adapter/common/src/test/scala/uk/ac/wellcome/sierra_adapter/model/SierraItemRecordTest.scala
@@ -1,10 +1,10 @@
 package uk.ac.wellcome.sierra_adapter.model
 
 import java.time.Instant
-
 import org.scalatest.funspec.AnyFunSpec
 import org.scalatest.matchers.should.Matchers
 import uk.ac.wellcome.json.JsonUtil._
+import weco.catalogue.sierra_adapter.generators.SierraGenerators
 
 class SierraItemRecordTest
     extends AnyFunSpec
