+++ sierra_adapter/common/src/test/scala/uk/ac/wellcome/sierra_adapter/model/SierraTransformableTest.scala
@@ -2,6 +2,7 @@ package uk.ac.wellcome.sierra_adapter.model
 
 import org.scalatest.funspec.AnyFunSpec
 import org.scalatest.matchers.should.Matchers
+import weco.catalogue.sierra_adapter.generators.SierraGenerators
 
 class SierraTransformableTest
     extends AnyFunSpec
