+++ sierra_adapter/sierra_linker/src/test/scala/weco/catalogue/sierra_linker/SierraLinkerFeatureTest.scala
@@ -8,6 +8,7 @@ import uk.ac.wellcome.messaging.memory.MemoryMessageSender
 import uk.ac.wellcome.sierra_adapter.model.Implicits._
 import uk.ac.wellcome.sierra_adapter.model._
 import uk.ac.wellcome.storage.store.memory.MemoryVersionedStore
+import weco.catalogue.sierra_adapter.generators.SierraGenerators
 import weco.catalogue.sierra_linker.fixtures.WorkerFixture
 import weco.catalogue.sierra_linker.models.Link
