+++ sierra_adapter/sierra_merger/src/test/scala/weco/catalogue/sierra_merger/SierraRecordMergerFeatureTest.scala
@@ -13,6 +13,7 @@ import uk.ac.wellcome.messaging.memory.MemoryMessageSender
 import uk.ac.wellcome.sierra_adapter.model._
 import uk.ac.wellcome.sierra_adapter.model.Implicits._
 import uk.ac.wellcome.storage.{Identified, Version}
+import weco.catalogue.sierra_adapter.generators.SierraGenerators
 import weco.catalogue.sierra_merger.fixtures.RecordMergerFixtures
 import weco.catalogue.sierra_merger.models.TransformableOps
 import weco.catalogue.sierra_merger.services.Worker
