+++ sierra_adapter/sierra_merger/src/test/scala/weco/catalogue/sierra_merger/models/TransformableOpsTest.scala
@@ -2,10 +2,8 @@ package weco.catalogue.sierra_merger.models
 
 import org.scalatest.funspec.AnyFunSpec
 import org.scalatest.matchers.should.Matchers
-import uk.ac.wellcome.sierra_adapter.model.{
-  SierraGenerators,
-  SierraTransformable
-}
+import uk.ac.wellcome.sierra_adapter.model.SierraTransformable
+import weco.catalogue.sierra_adapter.generators.SierraGenerators
 
 class TransformableOpsTest
     extends AnyFunSpec
