+++ sierra_adapter/sierra_merger/src/test/scala/weco/catalogue/sierra_merger/services/UpdaterTest.scala
@@ -4,7 +4,6 @@ import org.scalatest.{Assertion, EitherValues}
 import org.scalatest.funspec.AnyFunSpec
 import uk.ac.wellcome.sierra_adapter.model.Implicits._
 import uk.ac.wellcome.sierra_adapter.model.{
-  SierraGenerators,
   SierraItemRecord,
   SierraTransformable,
   TypedSierraRecordNumber
@@ -20,6 +19,7 @@ import uk.ac.wellcome.storage.store.{
   VersionedHybridStore
 }
 import uk.ac.wellcome.storage.{StoreWriteError, UpdateWriteError, Version}
+import weco.catalogue.sierra_adapter.generators.SierraGenerators
 import weco.catalogue.sierra_merger.fixtures.RecordMergerFixtures
 import weco.catalogue.source_model.fixtures.SourceVHSFixture
 import weco.catalogue.source_model.store.SourceVHS
